requirements.txt
numpy
opencv-python
pyyaml
tqdm
torch
torchvision